% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_family_df.R
\name{compute_family_df}
\alias{compute_family_df}
\title{Compute a data.frame of family relations for a start URL}
\usage{
compute_family_df(start_url)
}
\arguments{
\item{start_url}{URL in a Geni.com page}
}
\value{
A data.frame with family relations
}
\description{
Compute a data.frame of family relations for a start URL
}
\examples{
compute_family_df('https://www.geni.com/people/Guillame-le-Blanc-du-Bec-Seigneur-du-Bec/6000000008289847966')
}
