% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_from_url.R
\name{get_info_from_url}
\alias{get_info_from_url}
\title{Helper to get the information of all relatives}
\usage{
get_info_from_url(url)
}
\arguments{
\item{url}{A geni.com URL}
}
\value{
data.frame with url, birth date, death date, pob, pod
}
\description{
Helper to get the information of all relatives
}
\examples{
get_info_from_url('https://www.geni.com/people/Lucas-Brouwer/6000000037332972722')
}
