% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lijst_elite_moes.R
\docType{data}
\name{lijst_elite_moes}
\alias{lijst_elite_moes}
\title{List of Elite Families by Moes (2012)}
\format{
An object of class \code{data.frame}.
}
\source{
\href{https://pure.rug.nl/ws/portalfiles/portal/6550708/ThesisMoes.pdf}{RUG Portal}
}
\usage{
data(lijst_elite_moes)
}
\description{
A dataset from Moes (2012) contains the last names and prefixes
and frequencies of families considered to be part of the fiscal
and political elite in the Netherlands in the 19th and early 20th
centuries.
}
\examples{
data(lijst_elite_moes)
lijst_elite_moes
}
\references{
Moes (2012): Onder Aristocraten
(\href{https://pure.rug.nl/ws/portalfiles/portal/6550708/ThesisMoes.pdf}{RUG Portal})
}
\keyword{datasets}
