% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lijst_hoogst_aangeslagenen.R
\docType{data}
\name{lijst_hoogst_aangeslagenen}
\alias{lijst_hoogst_aangeslagenen}
\title{List of Highest Tax Payers by Moes (2012)}
\format{
An object of class \code{data.frame}.
}
\source{
\href{https://pure.rug.nl/ws/portalfiles/portal/6550708/ThesisMoes.pdf}{RUG Portal}
}
\usage{
data(lijst_elite_moes)
}
\description{
A dataset from Moes (2012) contains the last names, first names and prefixes
and frequencies of families considered to be part of the highest taxpayers
in the Netherlands in the 19th and early 20th centuries. It also contains the
place of birth and place of living at the moment of tax collection.
}
\examples{
data(lijst_hoogst_aangeslagenen)
lijst_hoogst_aangeslagenen
}
\references{
Moes (2012): Onder Aristocraten
(\href{https://pure.rug.nl/ws/portalfiles/portal/6550708/ThesisMoes.pdf}{RUG Portal})
}
\keyword{datasets}
