% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dates.R
\name{parse_dates}
\alias{parse_dates}
\title{Helper to parse dates from Geni.com}
\usage{
parse_dates(family)
}
\arguments{
\item{family}{data.frame with relations from geni.com}
}
\value{
data.frame with parsed dates
}
\description{
Helper to parse dates from Geni.com
}
\examples{
parse_dates(family)
}
