% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politician_diagnostics.R
\name{politician_diagnostics}
\alias{politician_diagnostics}
\title{Find out whether profile of wife's father or own father is in two prominent lists}
\usage{
politician_diagnostics(family, pdc_naam, lijst_ha, lijst_elite)
}
\arguments{
\item{family}{data.frame with family relations, e.g. father, spouse, mother}

\item{pdc_naam}{last name of subject in question him/herself}

\item{lijst_ha}{list higher aangeslagenen; can be NULL}

\item{lijst_elite}{list elite: can be NULL}
}
\value{
A data.frame with surnames, count siblings and counts on the two lists
}
\description{
Find out whether profile of wife's father or own father is in two prominent lists
}
\examples{
politician_diagnostics(family, "Jansen", lijst_hoogst_aangeslagenen, lijst_elite_moes)
}
