% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_moes_data.R
\name{scrape_moes_data}
\alias{scrape_moes_data}
\title{Scrape Moes Data}
\usage{
scrape_moes_data()
}
\value{
A list of df's with the Moes data
}
\description{
Scrape Moes Data
}
\examples{
dfs <- scrape_moes_data()
}
